macho: unblock stage2 on 32bit platforms (#7632)
* macho: unblock stage2 on 32bit platforms

Unblocks compilation of stage2 on 32bit platforms, and fixes #7630.

* Use libstd convention: reads - usize, writes - u64

+++ src/link/MachO/DebugSymbols.zig
@@ -747,7 +747,7 @@ fn updateDwarfSegment(self: *DebugSymbols) void {
 fn writeLoadCommands(self: *DebugSymbols, allocator: *Allocator) !void {
     if (!self.load_commands_dirty) return;
 
-    var sizeofcmds: usize = 0;
+    var sizeofcmds: u32 = 0;
     for (self.load_commands.items) |lc| {
         sizeofcmds += lc.cmdsize();
     }
